Update data for current weather displays
Both the top left and right now weather displays were using data from the day overview data from the API. This data did not represent the current weather accurately. The icons from this section seemed to indicate a guess on the day's weather as a whole, rather than the current weather. Instead, the data displayed will be from the hourly weather data.