Update carousel sliders to return correctly to base position
Earlier, a carousel could move to the right fine, but could not return all the way back to the starting position when returning to the left. Quite simply we needed to allow the left clicks to permit the position to be greater than or equal to zero, rather than simply allowing only greater than zero positions.