Add the ability to toggle the units being displayed
Now the user can change the units from celsius to fahrenheit using the top toggle button. If the user wants to store a preference for units, this method will probably need to change. In the initial element construction, the values are hardcoded to be fahrenheit, rather than reading the toggle state.