Fix rounding errors when changing units
Previously, since everything is rounded, when doing a unit change there would be different number after the unit toggle than what previously was there for the same unit. For example, it would initially display 39 degrees fahrenheit. Then after toggling to celsius and back to fahrenheit, it would display 38 degrees fahrenheit. This from the rounding that occurs. To solve this, I pre-rounded all the temperatures by converting them to celsius and back after the API call. This pre-rounded values is what is stored in the filtered data object that is used to update the display.