Update initial weather display logic
Now the weather will display any previously viewed location via local Storage. This allows the user to refresh the page and continue to view the location's weather they were looking at prior to the refresh. Additionally, on initial load, if the session storage is empty it will display the user's first saved location. Otherwise, if the user does not have any saved locations, it will display the weather for New York City.